+++ presentation/src/main/java/com/example/presentation/ui/interview/InterviewFragment.kt
@@ -1,4 +1,4 @@
-package com.example.presentation.analysis
+package com.example.presentation.ui.interview
 
 import androidx.lifecycle.ViewModelProvider
 import android.os.Bundle
@@ -8,24 +8,24 @@ import android.view.View
 import android.view.ViewGroup
 import com.example.presentation.R
 
-class AnalysisFragment : Fragment() {
+class InterviewFragment : Fragment() {
 
     companion object {
-        fun newInstance() = AnalysisFragment()
+        fun newInstance() = InterviewFragment()
     }
 
-    private lateinit var viewModel: AnalysisViewModel
+    private lateinit var viewModel: InterviewViewModel
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.fragment_analysis, container, false)
+        return inflater.inflate(R.layout.fragment_interview, container, false)
     }
 
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
-        viewModel = ViewModelProvider(this).get(AnalysisViewModel::class.java)
+        viewModel = ViewModelProvider(this).get(InterviewViewModel::class.java)
         // TODO: Use the ViewModel
     }
