+++ presentation/src/main/java/com/example/presentation/ui/interview/InterviewViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.presentation.interview
+package com.example.presentation.ui.interview
 
 import androidx.lifecycle.ViewModel
