+++ presentation/src/main/java/com/example/presentation/ui/login/LoginFragment.kt
@@ -1,5 +1,6 @@
-package com.example.presentation.signup
+package com.example.presentation.ui.login
 
+import android.content.Intent
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
@@ -7,36 +8,43 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import androidx.navigation.fragment.findNavController
+import com.example.presentation.MainActivity
 import com.example.presentation.R
-import com.example.presentation.databinding.FragmentSignUpBinding
+import com.example.presentation.databinding.FragmentLoginBinding
 
-class SignUpFragment : Fragment() {
+class LoginFragment : Fragment() {
 
-    private var _binding: FragmentSignUpBinding? = null
-    private val binding: FragmentSignUpBinding
+    private var _binding: FragmentLoginBinding? = null
+    private val binding: FragmentLoginBinding
         get() = _binding!!
 
-    private lateinit var viewModel: SignUpViewModel
+    private lateinit var viewModel: LoginViewModel
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding = FragmentSignUpBinding.inflate(layoutInflater)
+        _binding = FragmentLoginBinding.inflate(layoutInflater)
 
         return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+        val intent = Intent(activity, MainActivity::class.java)
 
-        binding.btnNext.setOnClickListener {
-            findNavController().navigate(R.id.action_signUpFragment_to_jobSkillFragment)
+        binding.btnLogin.setOnClickListener {
+            startActivity(intent)
+        }
+
+        binding.tvSignUp.setOnClickListener {
+            findNavController().navigate(R.id.action_loginFragment_to_signUpFragment)
         }
     }
+
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
-        viewModel = ViewModelProvider(this).get(SignUpViewModel::class.java)
+        viewModel = ViewModelProvider(this).get(LoginViewModel::class.java)
         // TODO: Use the ViewModel
     }
