+++ presentation/src/main/java/com/example/presentation/ui/login/LoginViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.presentation.login
+package com.example.presentation.ui.login
 
 import androidx.lifecycle.ViewModel
