+++ presentation/src/main/java/com/example/presentation/ui/signup/SignUpFragment.kt
@@ -1,6 +1,5 @@
-package com.example.presentation.login
+package com.example.presentation.ui.signup
 
-import android.content.Intent
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
@@ -8,43 +7,36 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import androidx.navigation.fragment.findNavController
-import com.example.presentation.MainActivity
 import com.example.presentation.R
-import com.example.presentation.databinding.FragmentLoginBinding
+import com.example.presentation.databinding.FragmentSignUpBinding
 
-class LoginFragment : Fragment() {
+class SignUpFragment : Fragment() {
 
-    private var _binding: FragmentLoginBinding? = null
-    private val binding: FragmentLoginBinding
+    private var _binding: FragmentSignUpBinding? = null
+    private val binding: FragmentSignUpBinding
         get() = _binding!!
 
-    private lateinit var viewModel: LoginViewModel
+    private lateinit var viewModel: SignUpViewModel
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding = FragmentLoginBinding.inflate(layoutInflater)
+        _binding = FragmentSignUpBinding.inflate(layoutInflater)
 
         return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        val intent = Intent(activity, MainActivity::class.java)
 
-        binding.btnLogin.setOnClickListener {
-            startActivity(intent)
-        }
-
-        binding.tvSignUp.setOnClickListener {
-            findNavController().navigate(R.id.action_loginFragment_to_signUpFragment)
+        binding.btnNext.setOnClickListener {
+            findNavController().navigate(R.id.action_signUpFragment_to_jobSkillFragment)
         }
     }
-
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
-        viewModel = ViewModelProvider(this).get(LoginViewModel::class.java)
+        viewModel = ViewModelProvider(this).get(SignUpViewModel::class.java)
         // TODO: Use the ViewModel
     }
