+++ presentation/src/main/res/layout/fragment_analysis.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".analysis.AnalysisFragment">
+    tools:context=".ui.analysis.AnalysisFragment">
 
     <TextView
         android:layout_width="wrap_content"
