+++ presentation/src/main/res/layout/fragment_interview.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".interview.InterviewFragment">
+    tools:context=".ui.interview.InterviewFragment">
 
     <TextView
         android:layout_width="wrap_content"
