+++ presentation/src/main/res/layout/fragment_job_skill.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".jobskill.JobSkillFragment">
+    tools:context=".ui.jobskill.JobSkillFragment">
 
     <androidx.constraintlayout.widget.Guideline
         android:id="@+id/guideline_start"
