+++ presentation/src/main/res/layout/fragment_my_page.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".mypage.MyPageFragment">
+    tools:context=".ui.mypage.MyPageFragment">
 
     <TextView
         android:layout_width="wrap_content"
