+++ presentation/src/main/res/layout/fragment_sign_up.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".signup.SignUpFragment">
+    tools:context=".ui.signup.SignUpFragment">
 
     <androidx.constraintlayout.widget.Guideline
         android:id="@+id/guideline_start"
