+++ presentation/src/main/res/navigation/nav_graph.xml
@@ -7,17 +7,17 @@
 
     <fragment
         android:id="@+id/analysisFragment"
-        android:name="com.example.presentation.analysis.AnalysisFragment"
+        android:name="com.example.presentation.ui.analysis.AnalysisFragment"
         android:label="fragment_analysis"
         tools:layout="@layout/fragment_analysis" />
     <fragment
         android:id="@+id/interviewFragment"
-        android:name="com.example.presentation.interview.InterviewFragment"
+        android:name="com.example.presentation.ui.interview.InterviewFragment"
         android:label="fragment_interview"
         tools:layout="@layout/fragment_interview" />
     <fragment
         android:id="@+id/myPageFragment"
-        android:name="com.example.presentation.mypage.MyPageFragment"
+        android:name="com.example.presentation.ui.mypage.MyPageFragment"
         android:label="fragment_my_page"
         tools:layout="@layout/fragment_my_page" />
 </navigation>
